requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
bidict==0.23.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
decorator==4.4.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib @ file:///C:/Users/mirac/Desktop/Tutorials/Files/Projects/Real-time-Face-Recognition-Project-main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
joblib==1.4.2
keras==3.4.1
keyboard==0.13.5
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
moviepy==1.0.3
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.3
psutil==6.0.0
psycopg2==2.9.9
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytweening==1.2.0
pytz==2024.1
pywin32==306
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.7.1
scapy==2.5.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
waitress==3.0.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
winshell==0.6
wrapt==1.16.0
wsproto==1.2.0
